requirements.txt
dash==2.17.1
pandas==1.5.2
numpy==1.26.4
dash-bootstrap-components==1.6.0
dash-ag-grid==31.2.0
rdkit==2024.03.5
pubchempy==1.0.4
gunicorn
python-dotenv==1.0.1 
google-ai-generativelanguage==0.6.9
google-generativeai==0.8.1
# pydeseq2==0.4.12
# polars==1.12.0

